+++ pootle/apps/pootle_store/util.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright (C) Pootle contributors.
+# Copyright (C) Zing contributors.
 #
-# This file is a part of the Pootle project. It is distributed under the GPL3
+# This file is a part of the Zing project. It is distributed under the GPL3
 # or later license. See the LICENSE file for a copy of the license and the
 # AUTHORS file for copyright and authorship information.
 
@@ -87,7 +88,3 @@ def get_change_str(changes):
 
 def get_state_name(code, default="untranslated"):
     return STATES_NAMES.get(code, default)
-
-
-def vfolders_installed():
-    return "virtualfolder" in settings.INSTALLED_APPS