+++ pootle/apps/pootle_translationproject/views.py
@@ -1,22 +1,22 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright (C) Pootle contributors.
+# Copyright (C) Zing contributors.
 #
-# This file is a part of the Pootle project. It is distributed under the GPL3
+# This file is a part of the Zing project. It is distributed under the GPL3
 # or later license. See the LICENSE file for a copy of the license and the
 # AUTHORS file for copyright and authorship information.
 
 import functools
 
-from django.conf import settings
 from django.core.urlresolvers import resolve, reverse
 from django.http import Http404
 from django.shortcuts import get_object_or_404, redirect
 from django.utils.functional import cached_property
 from django.utils.lru_cache import lru_cache
 
-from pootle.core.browser import (
-    get_parent, get_table_headings, make_directory_item, make_store_item)
+from pootle.core.browser import (get_parent, make_directory_item,
+                                 make_store_item)
 from pootle.core.decorators import get_path_obj, permission_required
 from pootle.core.helpers import get_sidebar_announcements_context
 from pootle.core.views import (
@@ -247,86 +247,21 @@ class TPBrowseView(TPDirectoryMixin, TPBrowseBaseView):
 
     @cached_property
     def items(self):
-        if 'virtualfolder' in settings.INSTALLED_APPS:
-            from virtualfolder.helpers import vftis_for_child_dirs
-            dirs_with_vfolders = set(
-                vftis_for_child_dirs(self.object).values_list(
-                    "directory__pk", flat=True))
-        else:
-            dirs_with_vfolders = []
         directories = [
-            make_directory_item(
-                child,
-                **(dict(sort="priority")
-                   if child.pk in dirs_with_vfolders
-                   else {}))
+            make_directory_item(child)
             for child in self.object.children
-            if isinstance(child, Directory)]
+            if isinstance(child, Directory)
+        ]
         stores = [
             make_store_item(child)
             for child in self.object.children
-            if isinstance(child, Store)]
+            if isinstance(child, Store)
+        ]
         return directories + stores
 
     @cached_property
-    def has_vfolders(self):
-        return self.object.has_vfolders
-
-    @cached_property
-    def vfolders(self):
-        from virtualfolder.helpers import make_vfolder_treeitem_dict
-        vftis = self.object.vf_treeitems
-        if not self.has_admin_access:
-            vftis = vftis.filter(vfolder__is_public=True)
-        return [
-            make_vfolder_treeitem_dict(vfolder_treeitem)
-            for vfolder_treeitem
-            in vftis.order_by('-vfolder__priority').select_related("vfolder")
-            if (self.has_admin_access
-                or vfolder_treeitem.is_visible)]
-
-    @cached_property
-    def vfolder_data(self):
-        ctx = {}
-        if 'virtualfolder' not in settings.INSTALLED_APPS:
-            return {}
-        if len(self.vfolders) > 0:
-            table_fields = [
-                'name', 'priority', 'progress', 'total',
-                'need-translation', 'suggestions', 'critical',
-                'last-updated', 'activity']
-            ctx.update({
-                'vfolders': {
-                    'id': 'vfolders',
-                    'fields': table_fields,
-                    'headings': get_table_headings(table_fields),
-                    'items': self.vfolders}})
-        return ctx
-
-    @cached_property
-    def vfolder_stats(self):
-        if 'virtualfolder' not in settings.INSTALLED_APPS:
-            return {}
-        stats = {"vfolders": {}}
-        for vfolder_treeitem in self.vfolders or []:
-            stats['vfolders'][
-                vfolder_treeitem['code']] = vfolder_treeitem["stats"]
-            del vfolder_treeitem["stats"]
-        return stats
-
-    @cached_property
     def stats(self):
-        stats = self.vfolder_stats
-        if stats and stats["vfolders"]:
-            stats.update(self.object.get_stats())
-        else:
-            stats = self.object.get_stats()
-        return stats
-
-    def get_context_data(self, *args, **kwargs):
-        ctx = super(TPBrowseView, self).get_context_data(*args, **kwargs)
-        ctx.update(self.vfolder_data)
-        return ctx
+        return self.object.get_stats()
 
 
 class TPTranslateBaseView(PootleTranslateView):
@@ -346,48 +281,13 @@ class TPTranslateView(TPDirectoryMixin, TPTranslateBaseView):
     def request_path(self):
         return "/%(language_code)s/%(project_code)s/%(dir_path)s" % self.kwargs
 
-    @cached_property
-    def extracted_path(self):
-        if 'virtualfolder' not in settings.INSTALLED_APPS:
-            return None, self.request_path
-
-        from virtualfolder.helpers import extract_vfolder_from_path
-        from virtualfolder.models import VirtualFolderTreeItem
-        return extract_vfolder_from_path(
-            self.request_path,
-            vfti=VirtualFolderTreeItem.objects.select_related(
-                "directory", "vfolder"))
-
-    @property
-    def display_vfolder_priority(self):
-        if 'virtualfolder' not in settings.INSTALLED_APPS:
-            return False
-        vfolder = self.extracted_path[0]
-        if vfolder:
-            return False
-        return self.object.has_vfolders
-
     @property
     def resource_path(self):
-        vfolder = self.extracted_path[0]
-        path = ""
-        if vfolder:
-            path = "%s/" % vfolder.name
-        return (
-            "%s%s"
-            % (path,
-               self.object.pootle_path.replace(self.ctx_path, "")))
+        return self.object.pootle_path.replace(self.ctx_path, '')
 
     @property
     def path(self):
-        return self.extracted_path[1]
-
-    @property
-    def vfolder_pk(self):
-        vfolder = self.extracted_path[0]
-        if vfolder:
-            return vfolder.pk
-        return ""
+        return self.request_path
 
 
 class TPTranslateStoreView(TPStoreMixin, TPTranslateBaseView):