+++ pytest_pootle/env.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright (C) Pootle contributors.
+# Copyright (C) Zing contributors.
 #
-# This file is a part of the Pootle project. It is distributed under the GPL3
+# This file is a part of the Zing project. It is distributed under the GPL3
 # or later license. See the LICENSE file for a copy of the license and the
 # AUTHORS file for copyright and authorship information.
 
@@ -21,7 +22,8 @@ class PootleTestEnv(object):
         "languages", "site_matrix", "system_users", "permissions",
         "site_permissions", "tps",
         "disabled_project", "subdirs", "submissions", "announcements",
-        "terminology", "fs", "vfolders", "complex_po")
+        "terminology", "fs", "complex_po",
+    )
 
     def setup(self, **kwargs):
         for method in self.methods:
@@ -465,52 +467,3 @@ class PootleTestEnv(object):
                 unit, old_target, old_state,
                 current_time, member, SubmissionTypes.NORMAL)
             unit.save()
-
-    def setup_vfolders(self):
-        from pytest_pootle.factories import VirtualFolderDBFactory
-
-        from django.db import connection
-        from django.apps import apps
-
-        from pootle.core.utils.db import set_mysql_collation_for_column
-
-        cursor = connection.cursor()
-
-        # VirtualFolderTreeItem
-        set_mysql_collation_for_column(
-            apps,
-            cursor,
-            "virtualfolder.VirtualFolderTreeItem",
-            "pootle_path",
-            "utf8_bin",
-            "varchar(255)")
-
-        # VirtualFolder
-        set_mysql_collation_for_column(
-            apps,
-            cursor,
-            "virtualfolder.VirtualFolder",
-            "name",
-            "utf8_bin",
-            "varchar(70)")
-        set_mysql_collation_for_column(
-            apps,
-            cursor,
-            "virtualfolder.VirtualFolder",
-            "location",
-            "utf8_bin",
-            "varchar(255)")
-
-        VirtualFolderDBFactory(filter_rules="store0.po")
-        VirtualFolderDBFactory(filter_rules="store1.po")
-        VirtualFolderDBFactory(
-            location='/{LANG}/project0/',
-            is_public=False,
-            filter_rules="store0.po")
-        VirtualFolderDBFactory(
-            location='/{LANG}/project0/',
-            is_public=False,
-            filter_rules="store1.po")
-        VirtualFolderDBFactory(
-            location='/language0/project0/',
-            filter_rules="subdir0/store4.po")