+++ tests/commands/minus_minus_help.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright (C) Pootle contributors.
+# Copyright (C) Zing contributors.
 #
-# This file is a part of the Pootle project. It is distributed under the GPL3
+# This file is a part of the Zing project. It is distributed under the GPL3
 # or later license. See the LICENSE file for a copy of the license and the
 # AUTHORS file for copyright and authorship information.
 
@@ -11,7 +12,7 @@ import pytest
 from django.core.management import call_command, get_commands
 
 CORE_APPS_WITH_COMMANDS = (
-    'accounts', 'pootle_app', 'virtualfolder',
+    'accounts', 'pootle_app',
 )