+++ tests/core/mixins/treeitem.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright (C) Pootle contributors.
+# Copyright (C) Zing contributors.
 #
-# This file is a part of the Pootle project. It is distributed under the GPL3
+# This file is a part of the Zing project. It is distributed under the GPL3
 # or later license. See the LICENSE file for a copy of the license and the
 # AUTHORS file for copyright and authorship information.
 
@@ -34,8 +35,7 @@ def test_get_children(project0, language0):
 
 
 @pytest.mark.django_db
-def test_get_parents(po_directory, project0, language0, tp0, store0, subdir0,
-                     no_vfolders):
+def test_get_parents(po_directory, project0, language0, tp0, store0, subdir0):
     """Ensure that retrieved parent objects have a correct type."""
 
     subdir_store = subdir0.child_stores.first()