+++ CHANGES.md
@@ -8,6 +8,7 @@ Zing Changelog
 * Django 1.9 support.
 * Cleaned up features out of the scope of the project:
   * Upload/download of user files.
+  * Virtual folders.
 
 
 Pre 0.1.0